rag-chatbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere>=4.0.0
qdrant-client>=1.7.0
pypdf==3.17.4
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.0
pdfplumber>=0.10.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
requests>=2.30.0
python-multipart==0.0.6
pyjwt>=2.8.0
cryptography>=41.0.0